Makefile: Improve automatic dependency mechanism by eliminating the need to manually specify the libraries and object files that an executable needs to be linked with. Instead, it is sufficient to list the libraries that an object file needs to be linked with the executables that includes its header file.
This is very beneficial because it is easier to list the libraries
that an object file needs during the development of the object file
itself rather than having to discover what libraries an executable
will need.

The improved mechanism relies on detecting the header files that an
executable depends on. For example, if some of the header files are
detected in a list of header files needing certain libraries, the
libraries will be added into LDFLAGS of the executable. As another
example, if some header files have the corresponding source files, the
object files of the source files will be added as the prerequisites of
the executable.